Tidy tests, their data and their enabling
