+++ tests/pipelex/plugin/test_openai_image_gen.py
@@ -6,7 +6,7 @@ from pipelex.plugin.openai.openai_factory import OpenAIFactory
 from pipelex.tools.misc.base_64_utils import save_base64_to_binary_file
 from pipelex.tools.misc.file_utils import ensure_path, get_incremental_file_path
 from tests.conftest import TEST_OUTPUTS_DIR
-from tests.pipelex.cogt.test_data import IMGGTestCases
+from tests.pipelex.test_data import IMGGTestCases
 
 
 @pytest.mark.imgg
